+++ examples/apps/web-list/src/product-list/product-list.js
@@ -1,4 +1,4 @@
-import { html, list } from "@inglorious/lit"
+import { html, list } from "@inglorious/web"
 
 const DIVISOR = 2