+++ examples/apps/web-router/src/store/types/user-detail.js
@@ -1,4 +1,4 @@
-import { html } from "@inglorious/lit"
+import { html } from "@inglorious/web"
 
 const BACK = -1