+++ examples/apps/web-todomvc/src/store/index.js
@@ -1,4 +1,4 @@
-import { createStore } from "@inglorious/lit"
+import { createStore } from "@inglorious/web"
 
 import { app } from "../app"
 import { footer } from "../footer/footer"