+++ examples/apps/web-todomvc/src/store/middlewares.js
@@ -1,5 +1,5 @@
 // import { multiplayerMiddleware } from "@inglorious/store/client/multiplayer-middleware"
-import { createDevtools } from "@inglorious/lit"
+import { createDevtools } from "@inglorious/web"
 
 export const middlewares = [
   // multiplayerMiddleware({ blacklist: ["inputChange", "filterClick"] }),